requirements.txt
fyers-apiv3
numpy
pandas
pyotp
requests
python-dotenv
matplotlib